// SIMEM-ANG-ADMIN/src/app/store/model/destination-column/destination.column.model.ts
export interface DestinationColumnModel {
  idColumnaDestino: string;
  nombreColumnaDestino: string | null;
  tipoDato: string | null;
  atributoVariable: string | null;
  variableId: number | null;
  estado: boolean | null;
  descripcion: string | null;
  fechaCreacion: Date;
  fechaActualizacion: Date | null;

}